wordgame_bot/requirements.txt
psycopg2-binary
discord.py==1.7.3
python-dotenv==0.19.2
